feat: fix unmatched ICURVE interface

+++ contracts/6/protocol/interfaces/ICurveBase.sol
@@ -19,36 +19,36 @@ interface ICurveBase {
 
   function remove_liquidity_one_coin(
     uint256 _token_amount,
-    uint128 i,
+    int128 i,
     uint256 min_amount
   ) external;
 
   function remove_liquidity_imbalance(uint256[4] calldata uamounts, uint256 max_burn_amount) external;
 
   function exchange(
-    uint128 i,
-    uint128 j,
+    int128 i,
+    int128 j,
     uint256 dx,
     uint256 min_dy
   ) external;
 
   function calc_withdraw_one_coin(
     uint256 _token_amount,
-    uint128 i,
+    int128 i,
     uint256 min_amount
   ) external view returns (uint256);
 
   function calc_token_amount() external view returns (uint256);
 
-  function coins(int128 i) external view returns (address);
+  function coins(uint256 i) external view returns (address);
 
   function fee() external view returns (uint256);
 
   function A() external view returns (uint256);
 
   function get_dy(
-    uint128 i,
-    uint128 j,
+    int128 i,
+    int128 j,
     uint256 dx
   ) external view returns (uint256);
